% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IC.var.R
\name{IC.var}
\alias{IC.var}
\title{Confidence Interval V}
\usage{
IC.var(x, n = length(x), conf.level = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{A vector or the value of the sample variance}

\item{n}{The size of the sample; it's not necessary to enter it when x is a vector}

\item{conf.level}{The confidence level; by default it's 0.95}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds; its value by default is FALSE}
}
\description{
Ths function computes a confidence interval for the variance of a normal random variable.
The result is a data frame with variables
variance (the sample variance),
size (the sample size),
lower and upper (the lower and upper confidence interval limits)
conf.level the confidence level
}
\examples{
IC.var
}
\keyword{confidence}
\keyword{interval}

