% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICT.exact.R
\name{ICT.exact}
\alias{ICT.exact}
\title{Confidence Interval II}
\usage{
ICT.exact(x, sdm = sd(x), n = length(x), conf.level = 0.95,
  na.rm = FALSE)
}
\arguments{
\item{x}{a vector or the value of the sample mean}

\item{sdm}{the sample standard deviation; it's not necessary to enter it when x is a vector}

\item{n}{the size of the sample; it's not necessary to enter it when x is a vector}

\item{conf.level}{the confidence level; by default it's 0.95}
}
\description{
Ths function computes a confidence interval for the mean of a normal random variable with unknown standard deviation.
The result is a data frame with variables
mean (the sample mean),
size (the sample size),
lower and upper (the lower and upper confidence interval limits)
conf.level the confidence level
}
\examples{
ICT.exact()
}
\keyword{confidence}
\keyword{interval}

