% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IC.approx.R
\name{ICZ.approx}
\alias{ICZ.approx}
\title{Confidence Interval III}
\usage{
ICZ.approx(x, sdm = sd(x), n = length(x), conf.level = 0.95,
  na.rm = FALSE)
}
\arguments{
\item{x}{A vector or the value of the sample mean}

\item{n}{The size of the sample; it's not necessary to enter it when x is a vector}

\item{conf.level}{The confidence level; by default it's 0.95}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds; its value by default is FALSE}

\item{sigma}{The standard deviation of the random variable}
}
\description{
Ths function computes an approximate confidence interval for the mean of a random variable with unknown standard deviation
when the sample is large.
The result is a data frame with variables
mean (the sample mean),
size (the sample size),
lower and upper (the lower and upper confidence interval limits)
conf.level the confidence level
}
\examples{
ICZ.approx()
}
\keyword{confidence}
\keyword{interval}

