% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICZ.exact.R
\name{ICZ.exact}
\alias{ICZ.exact}
\title{Confidence Interval I}
\usage{
ICZ.exact(x, sigma, n = length(x), conf.level = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or the value of the sample mean}

\item{sigma}{The standard deviation of the random variable}

\item{n}{The size of the sample; it should not be entered when x is a vector}

\item{conf.level}{The confidence level; by default it's 0.95}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds; its value by default is FALSE}
}
\description{
This function computes a confidence interval for the mean of a normal random variable with known standard deviation.
}
\details{
The result is a data frame with variables
mean (the sample mean),
size (the sample size),
lower and upper (the lower and upper confidence interval limits)
conf.level (the confidence level)
}
\examples{
ICZ.exact(0.5,1.8,45,0.9)
}
\keyword{confidence}
\keyword{interval}

