% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IC.approx.R
\name{ICZ.p}
\alias{ICZ.p}
\title{Confidence Interval IV}
\usage{
ICZ.p(x, n = length(x), conf.level = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of 0s and 1s or the number of 1s in the sample}

\item{n}{The size of the sample; it's not necessary to enter it when x is a vector}

\item{conf.level}{The confidence level; by default it's 0.95}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds; its value by default is FALSE}
}
\description{
Ths function computes an approximate confidence interval for the success probability of a Bernoulli variable
using Laplace formula; the sample, the number of successes and the number of failures must be large to be valid.
The result is a data frame with variables
proportion (the sample proportion),
size (the sample size),
lower and upper (the lower and upper confidence interval limits)
conf.level the confidence level
}
\examples{
ICZ.p()
}
\keyword{confidence}
\keyword{interval}

