% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NMin.R
\name{NMin.mu}
\alias{NMin.mu}
\title{Sample size}
\usage{
NMin.mu(A, sigma, conf.level = 0.95)
}
\arguments{
\item{A}{the target range}

\item{sigma}{the standard deviation of the random variable}

\item{conf.level}{the confidence level; by default it's 0.95}
}
\description{
This function computes the smallest necessary size of a sample of a normal population with lnown standard deviation
in order to obtain a confidence interval of a maximum target range
}
\examples{
NMin.mu()
}
\keyword{Size}

