% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NMin.R
\name{NMin.p}
\alias{NMin.p}
\title{Sample size}
\usage{
NMin.p(A, conf.level = 0.95)
}
\arguments{
\item{A}{the target range}

\item{conf.level}{the confidence level; by default it's 0.95}
}
\description{
This function computes the smallest necessary size of a sample of a Bernoulli population
in order to obtain a confidence interval of a maximum given range in the worst of cases
}
\examples{
NMin.p()
}
\keyword{Size}

